do not allocate 65535 size buffers every time
do not waste memory with the largest possible sized buffer for every
single operation. Break the free lists out so we have one of those still
for stream oriented connections, but have a smaller one for the udp
packets. Also allow this size to be administratively set so 8192 is just
a default. If this changes at runtime, the pool will adapt (at least
up). MDNS gets its own special pool with its well defined packet size.

Usage of EDNS is still a little clumsy.

+++ ns/server.go
@@ -59,6 +59,7 @@ func NewServer(
 
 // Serve runs a unicast server until the context is canceled.
 // It is safe and possible, although not necessarily beneficial, to have multiple Serve routines on the same Server instance
+// if the underlying connection is packet based.
 func (s *Server) Serve(ctx context.Context) error {
 	if s.conn == nil {
 		return ErrNoConnection
@@ -150,7 +151,7 @@ func messageSize(conn dnsconn.Conn, msg *dns.Message) int {
 				msgSize = dnsconn.MinMessageSize
 			}
 			msg.EDNS = &dns.Record{
-				H: dns.NewEDNSHeader(dnsconn.UDPMessageSize, 0, 0, 0),
+				H: dns.NewEDNSHeader(uint16(dnsconn.UDPMessageSize), 0, 0, 0),
 				D: &dns.EDNSRecord{},
 			}
 		}
